Red-team fixes: CNN buffer overflow, kernel wipe, BNN offset, attestation lifetime
A four-lens adversarial review of the remediation diff found four real
defects in the changed code. All fixed with load-bearing regression tests
(each verified to FAIL when the fix is reverted), full suite green under
ASan/UBSan.

1. CNN1D weight_scratch overflow (memory corruption). The engine stages all
   conv kernels (F*K bytes) into weight_scratch, sized OUTPUT*FLAT. That is
   NOT >= F*K in general (a large kernel with few output classes), so a valid
   .npz could compile into an out-of-bounds write — UBSan-confirmed
   (in=6,K=5,F=2,pool=2: index 5 into mnv_weight_t[2]). Root cause is the
   engine buffer sizing; the new CNN compiler is what made it reachable from
   a plain model file. Fixed by sizing weight_scratch to max(F*K, FLAT).
   Regression: cnn1d_bigkernel.

2. CNN1D kernel-wipe truncation (Law II). The post-use secure_zero of the
   staged kernels recomputed the length and cast it to uint16, leaving
   decrypted kernel weights resident in SRAM for blobs > 64 KB. Now uses the
   uint32 all_kernels length.

3. BNN neuron bit-offset wrap. bnn_dot_bits took a uint16 bit offset, so a
   single layer with in_sz*out_sz >= 65536 bits wrapped n*in_sz and read the
   wrong packed weights (garbage output, not rejection). Widened w_off/wi to
   uint32. Regression: bnn_bigoffset (256->264 layer).

4. Output-attestation lifetime asymmetry. The input-validation and
   confidence-check early returns bypassed the fail label, so they did NOT
   clear has_output_mac/output_mac — a rejected inference left the prior
   success's attestation live (inconsistent with the glitch/mismatch paths).
   Not a false-accept (the flag is never true while the MAC is zeroed), but
   the per-failure-type asymmetry contradicted Item 9's intent. Routed both
   early returns through fail for uniform invalidation. Regression:
   reject_clears_mac.

Also: CNN compiler now warns when conv_len isn't divisible by pool (silent
tail-drop). Documented a pre-existing, out-of-scope AVR issue surfaced by the
review — layer descriptors are read without pgm_read while emitted PROGMEM,
so they read garbage on real AVR (needs AVR-hardware validation to fix).

+++ src/arch/mnv_bnn.c
@@ -53,12 +53,14 @@
  * @param len     Number of binary values (in_sz).
  * @return Signed accumulator in range [-len, +len].
  */
-static int16_t bnn_dot_bits(const uint8_t *w, uint16_t w_off,
+static int16_t bnn_dot_bits(const uint8_t *w, uint32_t w_off,
                             const uint8_t *a, uint16_t len)
 {
     int16_t acc = 0;
     for (uint16_t i = 0; i < len; i++) {
-        uint16_t wi = (uint16_t)(w_off + i);
+        /* uint32: a single layer with in_sz*out_sz >= 65536 bits pushes the
+         * neuron bit offset past uint16 (would wrap and read wrong weights). */
+        uint32_t wi = w_off + (uint32_t)i;
         uint8_t  wb = (uint8_t)((w[wi >> 3] >> (wi & 7u)) & 1u);
         uint8_t  ab = (uint8_t)((a[i  >> 3] >> (i  & 7u)) & 1u);
         /* xor==0 -> agree -> +1 ; xor==1 -> disagree -> -1 (branchless) */
@@ -145,7 +147,7 @@ mnv_status_t mnv_bnn_forward(mnv_ctx_t          *ctx,
         for (uint16_t n = 0; n < out_sz; n++) {
             /* Weights for neuron n start at bit n*in_sz in the packed buffer. */
             int16_t acc = bnn_dot_bits((const uint8_t *)ctx->weight_scratch,
-                                       (uint16_t)(n * in_sz), packed_src, in_sz);
+                                       (uint32_t)n * (uint32_t)in_sz, packed_src, in_sz);
 
             /* Scale to Q8: acc ∈ [-in_sz, +in_sz], map to [-127, +127] */
             int16_t scaled = (int16_t)((int32_t)acc * 127 / (int16_t)in_sz);
